renamed redfish.* to pwr_redfish.* for consistent node naming
